Moving flow types back to individual files
