+++ lib/chrome-messages.js
@@ -2,7 +2,10 @@
 /*global chrome:false*/
 'use strict';
 
-import type {ChromeMessage} from './flowtypes';
+export type ChromeMessage = {
+    type: string;
+    body?: ?any; // should be Object | string; but in flow, Array is not Object
+};
 
 export function exists(): Promise {
     if (typeof chrome === 'undefined') {
