+++ lib/descriptor-stream.js
@@ -4,7 +4,17 @@
 import {EventEmitter} from './events';
 import PluginTransport from './transport/plugin';
 
-import type {Transport, DeviceDescriptor, DeviceDescriptorDiff} from './flowtypes';
+import type {Transport, DeviceDescriptor} from './transport';
+
+export type DeviceDescriptorDiff = {
+    connected: Array<DeviceDescriptor>,
+    disconnected: Array<DeviceDescriptor>,
+    changedSessions: Array<DeviceDescriptor>,
+    acquired: Array<DeviceDescriptor>,
+    released: Array<DeviceDescriptor>,
+    didUpdate: boolean,
+    descriptors: Array<DeviceDescriptor>
+};
 
 export default class DescriptorStream extends EventEmitter {
