+++ lib/device-list.js
@@ -7,7 +7,17 @@ import DescriptorStream from './descriptor-stream';
 import Device from './device';
 import UnacquiredDevice from './unacquired-device';
 
-import type {DeviceDescriptorDiff, Transport, DeviceDescriptor, DeviceListOptions} from './flowtypes';
+import type {DeviceDescriptorDiff} from './descriptor-stream';
+import type {
+    Transport,
+    DeviceDescriptor,
+} from './transport';
+
+export type DeviceListOptions = {
+    transport?: Transport;
+    configUrl?: string;
+    config?: string;
+};
 
 // a slight hack
 // this error string is hard-coded
