+++ lib/device.js
@@ -5,8 +5,13 @@ import semvercmp from 'semver-compare';
 
 import {EventEmitter} from './events';
 import Session from './session';
+
 import type DeviceList from './device-list';
-import type {Features, Transport, DeviceDescriptor} from './flowtypes';
+import type {Features} from './flowtypes';
+import type {
+    Transport,
+    DeviceDescriptor,
+} from './transport';
 
 // a slight hack
 // this error string is hard-coded
