+++ lib/flowtypes.js
@@ -1,33 +1,6 @@
 /* @flow */
 'use strict';
 
-export type MessageResponse<T> = {
-    type: string;
-    message: T; // in general, can be anything
-};
-
-export type DefaultMessageResponse = MessageResponse<Object>;
-
-export type DeviceDescriptor = {
-    path: string;
-    session: ?string;
-    product: number;
-    serialNumber: number;
-    vendor: number;
-};
-
-export type Transport = {
-    enumerate: (wait?: boolean, previous?: ?Array<DeviceDescriptor>) => Promise<Array<DeviceDescriptor>>;
-    acquire: (i: DeviceDescriptor, checkPrevious?: boolean) => Promise<{session: string}>;
-    supportsSync: boolean;
-    release: (s: string) => Promise;
-    releaseSync: (s: string) => void;
-    configure: (s: string) => Promise;
-    call: (s: string, type: string, msg: Object) => Promise<DefaultMessageResponse>;
-    callSync: (s: string, type: string, msg: Object) => DefaultMessageResponse;
-    version: string;
-};
-
 export type CoinType = {
     coin_name: string;
     coin_shortcut: string;
@@ -191,25 +164,6 @@ export type PublicKey = {
     xpub: string;
 };
 
-export type OutputInfo = {
-    path: Array<number>;
-    value: number;
-} | {
-    address: string;
-    value: number;
-};
-
-export type InputInfo = {
-    hash: Buffer;
-    index: number;
-    path?: Array<number>;
-};
-
-export type TxInfo = {
-    inputs: Array<InputInfo>;
-    outputs: Array<OutputInfo>;
-};
-
 // this is what Trezor asks for
 export type SignTxInfoToTrezor = {
     inputs: Array<TransactionInput>
@@ -223,46 +177,3 @@ export type SignTxInfoToTrezor = {
     inputs_cnt: number;
     outputs_cnt: number;
 };
-
-export type DeviceDescriptorDiff = {
-    connected: Array<DeviceDescriptor>,
-    disconnected: Array<DeviceDescriptor>,
-    changedSessions: Array<DeviceDescriptor>,
-    acquired: Array<DeviceDescriptor>,
-    released: Array<DeviceDescriptor>,
-    didUpdate: boolean,
-    descriptors: Array<DeviceDescriptor>
-};
-
-export type DeviceListOptions = {
-    transport?: Transport;
-    configUrl?: string;
-    config?: string;
-};
-
-export type BridgeInstaller = {
-    version: string;
-    url: string;
-    label: string;
-    platform: string | Array<string>;
-    preferred: boolean;
-};
-
-export type UdevInstaller = {
-    url: string;
-    label: string;
-    platform: string | Array<string>;
-    preferred: boolean;
-};
-
-export type ChromeMessage = {
-    type: string;
-    body?: ?any; // should be Object | string; but in flow, Array is not Object
-};
-
-export type RequestOptions = {
-    body?: any; // should be Object | string; but in flow, Array is not Object
-    url: string;
-    method?: 'POST' | 'GET';
-};
-
