+++ lib/http.js
@@ -3,7 +3,11 @@
 
 const _syncRequest = require('sync-request/browser.js');
 
-import type {RequestOptions} from './flowtypes';
+export type RequestOptions = {
+    body?: any; // should be Object | string; but in flow, Array is not Object
+    url: string;
+    method?: 'POST' | 'GET';
+};
 
 function contentType(body: any): string {
     if (typeof body === 'string') {
