+++ lib/index.js
@@ -16,4 +16,5 @@ export {default as DeviceList} from './device-list';
 export {default as http} from './http';
 export {installers, latestVersion, udevInstallers} from './installers';
 
-export type {BridgeInstaller, UdevInstaller, Transport} from './flowtypes';
+export type {BridgeInstaller, UdevInstaller} from './installers';
+export type {Transport} from './transport';
