+++ lib/installers.js
@@ -1,8 +1,6 @@
 /* @flow */
 'use strict';
 
-import type {UdevInstaller, BridgeInstaller} from './flowtypes';
-
 const DATA_DOMAIN = 'https://mytrezor.s3.amazonaws.com';
 
 type ProtoInstallerShort = {
@@ -16,6 +14,21 @@ type ProtoInstaller = {
     label: string;
     platform: string | Array<string>;
 }
+
+export type BridgeInstaller = {
+    version: string;
+    url: string;
+    label: string;
+    platform: string | Array<string>;
+    preferred: boolean;
+};
+
+export type UdevInstaller = {
+    url: string;
+    label: string;
+    platform: string | Array<string>;
+    preferred: boolean;
+};
 
 function fillInstallerUrl(installer: ProtoInstallerShort): ProtoInstaller {
     return {
