+++ lib/session.js
@@ -3,14 +3,26 @@
 
 import {EventEmitter} from './events';
 import * as bitcoin from 'bitcoinjs-lib';
-import * as types from './flowtypes';
 
 import * as hdnodeUtils from './utils/hdnode';
 import * as signTxHelper from './utils/signtx';
 import * as signBjsTxHelper from './utils/signbjstx';
 import {CallHelper} from './utils/call';
-
 import wrapLoadDevice from './utils/wrapLoadDevice';
+
+import * as types from './flowtypes';
+import type {TxInfo} from './utils/signbjstx';
+import type {
+    Transport,
+    DeviceDescriptor,
+} from './transport';
+
+export type MessageResponse<T> = {
+    type: string;
+    message: T; // in general, can be anything
+};
+
+export type DefaultMessageResponse = MessageResponse<Object>;
 
 //
 // Trezor device session handle. Acts as a event emitter.
@@ -27,13 +39,13 @@ import wrapLoadDevice from './utils/wrapLoadDevice';
 //  passphrase: callback(error, passphrase)
 //
 export default class Session extends EventEmitter {
-    _transport: types.Transport;
+    _transport: Transport;
     _sessionId: string;
-    _descriptor: types.DeviceDescriptor;
+    _descriptor: DeviceDescriptor;
     supportsSync: boolean;
     callHelper: CallHelper;
 
-    constructor(transport: types.Transport, sessionId: string, descriptor: types.DeviceDescriptor) {
+    constructor(transport: Transport, sessionId: string, descriptor: DeviceDescriptor) {
         super();
         this._transport = transport;
         this._sessionId = sessionId;
@@ -50,7 +62,7 @@ export default class Session extends EventEmitter {
         return this._descriptor.path;
     }
 
-    isDescriptor(descriptor: types.DeviceDescriptor): boolean {
+    isDescriptor(descriptor: DeviceDescriptor): boolean {
         return this._descriptor.path === descriptor.path;
     }
 
@@ -70,21 +82,21 @@ export default class Session extends EventEmitter {
         return this._transport.releaseSync(this._sessionId);
     }
 
-    initialize(): Promise<types.MessageResponse<types.Features>> {
+    initialize(): Promise<MessageResponse<types.Features>> {
         return this.typedCall('Initialize', 'Features');
     }
 
-    getFeatures(): Promise<types.MessageResponse<types.Features>> {
+    getFeatures(): Promise<MessageResponse<types.Features>> {
         return this.typedCall('GetFeatures', 'Features');
     }
 
-    getEntropy(size: number): Promise<types.MessageResponse<{bytes: string}>> {
+    getEntropy(size: number): Promise<MessageResponse<{bytes: string}>> {
         return this.typedCall('GetEntropy', 'Entropy', {
             size: size,
         });
     }
 
-    getAddress(address_n: Array<number>, coin: types.CoinType | string, show_display: ?boolean): Promise<types.MessageResponse<{
+    getAddress(address_n: Array<number>, coin: types.CoinType | string, show_display: ?boolean): Promise<MessageResponse<{
         address: string;
         path: Array<number>;
     }>> {
@@ -99,7 +111,7 @@ export default class Session extends EventEmitter {
         });
     }
 
-    getPublicKey(address_n: Array<number>): Promise<types.MessageResponse<types.PublicKey>> {
+    getPublicKey(address_n: Array<number>): Promise<MessageResponse<types.PublicKey>> {
         return this.typedCall('GetPublicKey', 'PublicKey', {
             address_n: address_n,
         }).then(res => {
@@ -108,64 +120,64 @@ export default class Session extends EventEmitter {
         });
     }
 
-    wipeDevice(): Promise<types.DefaultMessageResponse> {
+    wipeDevice(): Promise<DefaultMessageResponse> {
         return this.typedCall('WipeDevice', 'Success');
     }
 
-    resetDevice(settings: types.ResetDeviceSettings): Promise<types.DefaultMessageResponse> {
+    resetDevice(settings: types.ResetDeviceSettings): Promise<DefaultMessageResponse> {
         return this.typedCall('ResetDevice', 'Success', settings);
     }
 
     loadDevice(
         settings: types.LoadDeviceSettings,
         network: types.CoinType | string | bitcoin.Network
-    ): Promise<types.DefaultMessageResponse> {
+    ): Promise<DefaultMessageResponse> {
         return this.typedCall('LoadDevice', 'Success', wrapLoadDevice(settings, coinNetwork(network)));
     }
 
-    recoverDevice(settings: types.RecoverDeviceSettings): Promise<types.DefaultMessageResponse> {
+    recoverDevice(settings: types.RecoverDeviceSettings): Promise<DefaultMessageResponse> {
         return this.typedCall('RecoveryDevice', 'Success', {
             ...settings,
             enforce_wordlist: true,
         });
     }
 
-    applySettings(settings: types.ApplySettings): Promise<types.DefaultMessageResponse> {
+    applySettings(settings: types.ApplySettings): Promise<DefaultMessageResponse> {
         return this.typedCall('ApplySettings', 'Success', settings);
     }
 
-    clearSession(settings?: {}): Promise<types.DefaultMessageResponse> {
+    clearSession(settings?: {}): Promise<DefaultMessageResponse> {
         return this.typedCall('ClearSession', 'Success', settings);
     }
 
     // Blocks the browser thread, be careful.
-    clearSessionSync(settings?: {}): types.DefaultMessageResponse {
+    clearSessionSync(settings?: {}): DefaultMessageResponse {
         return this._callSync('ClearSession', settings);
     }
 
-    changePin(remove: ?boolean): Promise<types.DefaultMessageResponse> {
+    changePin(remove: ?boolean): Promise<DefaultMessageResponse> {
         return this.typedCall('ChangePin', 'Success', {
             remove: remove || false,
         });
     }
 
-    eraseFirmware(): Promise<types.DefaultMessageResponse> {
+    eraseFirmware(): Promise<DefaultMessageResponse> {
         return this.typedCall('FirmwareErase', 'Success');
     }
 
     // payload is in hexa
-    uploadFirmware(payload: string): Promise<types.DefaultMessageResponse> {
+    uploadFirmware(payload: string): Promise<DefaultMessageResponse> {
         return this.typedCall('FirmwareUpload', 'Success', {
             payload: payload,
         });
     }
 
-    updateFirmware(payload: string): Promise<types.DefaultMessageResponse> {
+    updateFirmware(payload: string): Promise<DefaultMessageResponse> {
         return this.eraseFirmware().then(() => this.uploadFirmware(payload));
     }
 
     // failure to verify returns rejecting promise
-    verifyMessage(address: string, signature: string, message: string): Promise<types.DefaultMessageResponse> {
+    verifyMessage(address: string, signature: string, message: string): Promise<DefaultMessageResponse> {
         return this.typedCall('VerifyMessage', 'Success', {
             address: address,
             signature: signature,
@@ -173,7 +185,7 @@ export default class Session extends EventEmitter {
         });
     }
 
-    signMessage(address_n: Array<number>, message: string, coin: types.CoinType | string): Promise<types.MessageResponse<types.MessageSignature>> {
+    signMessage(address_n: Array<number>, message: string, coin: types.CoinType | string): Promise<MessageResponse<types.MessageSignature>> {
         return this.typedCall('SignMessage', 'MessageSignature', {
             address_n: address_n,
             message: message,
@@ -185,7 +197,7 @@ export default class Session extends EventEmitter {
         identity: types.Identity,
         challenge_hidden: string,
         challenge_visual: string
-    ): Promise<types.MessageResponse<types.SignedIdentity>> {
+    ): Promise<MessageResponse<types.SignedIdentity>> {
         return this.typedCall('SignIdentity', 'SignedIdentity', {
             identity: identity,
             challenge_hidden: challenge_hidden,
@@ -201,7 +213,7 @@ export default class Session extends EventEmitter {
         ask_on_encrypt: ?boolean,
         ask_on_decrypt: ?boolean,
         iv: ?(string | Buffer) // in hexadecimal
-    ): Promise<types.MessageResponse<{value: string}>> {
+    ): Promise<MessageResponse<{value: string}>> {
         const valueString: string = (typeof value === 'string') ? value : value.toString('hex');
         const ivString: ?string = iv == null ? null : ((typeof iv === 'string') ? iv : iv.toString('hex'));
 
@@ -220,7 +232,7 @@ export default class Session extends EventEmitter {
         inputs: Array<types.TransactionInput>,
         outputs: Array<types.TransactionInput>,
         coin: string | types.CoinType
-    ): Promise<types.MessageResponse<{tx_size: number}>> {
+    ): Promise<MessageResponse<{tx_size: number}>> {
         return this.typedCall('EstimateTxSize', 'TxSize', {
             inputs_count: inputs.length,
             outputs_count: outputs.length,
@@ -233,12 +245,12 @@ export default class Session extends EventEmitter {
         outputs: Array<types.TransactionOutput>,
         txs: Array<types.RefTransaction>,
         coin: types.CoinType | string
-    ): Promise<types.MessageResponse<types.SignedTx>> {
+    ): Promise<MessageResponse<types.SignedTx>> {
         return signTxHelper.signTx(this, inputs, outputs, txs, coin);
     }
 
     signBjsTx(
-        info: types.TxInfo,
+        info: TxInfo,
         refTxs: Array<bitcoin.Transaction>,
         nodes: Array<bitcoin.HDNode>,
         coinName: string
@@ -246,18 +258,18 @@ export default class Session extends EventEmitter {
         return signBjsTxHelper.signBjsTx(info, refTxs, nodes, coinName);
     }
 
-    typedCall(type: string, resType: string, msg: Object = {}): Promise<types.DefaultMessageResponse> {
+    typedCall(type: string, resType: string, msg: Object = {}): Promise<DefaultMessageResponse> {
         return this.callHelper.typedCall(type, resType, msg);
     }
 
-    _call(type: string, msg: ?Object): Promise<types.DefaultMessageResponse> {
+    _call(type: string, msg: ?Object): Promise<DefaultMessageResponse> {
         return this.callHelper.call(type, msg);
     }
 
     // Sends a blocking message to an opened device. Be careful, the whole
     // tab thread gets blocked and does not respond. Also, we don't do any
     // timeouts here.
-    _callSync(type: string, msg: ?Object): types.DefaultMessageResponse {
+    _callSync(type: string, msg: ?Object): DefaultMessageResponse {
         return this.callHelper.callSync(type, msg);
     }
 
@@ -279,7 +291,7 @@ export default class Session extends EventEmitter {
         });
     }
 
-    changeLabel(label: string): Promise<types.DefaultMessageResponse> {
+    changeLabel(label: string): Promise<DefaultMessageResponse> {
         if (label.length > Session.LABEL_MAX_LENGTH) {
             label = label.slice(0, Session.LABEL_MAX_LENGTH);
         }
@@ -289,13 +301,13 @@ export default class Session extends EventEmitter {
         });
     }
 
-    togglePassphrase(enable: boolean): Promise<types.DefaultMessageResponse> {
+    togglePassphrase(enable: boolean): Promise<DefaultMessageResponse> {
         return this.applySettings({
             use_passphrase: enable,
         });
     }
 
-    changeHomescreen(hex: string): Promise<types.DefaultMessageResponse> {
+    changeHomescreen(hex: string): Promise<DefaultMessageResponse> {
         return this.applySettings({
             homescreen: hex,
         });
