+++ lib/transport/checks.js
@@ -1,10 +1,8 @@
 /* @flow */
 'use strict';
 
-import type {
-    DefaultMessageResponse,
-    DeviceDescriptor,
-} from '../flowtypes';
+import type {DeviceDescriptor} from '../transport';
+import type {DefaultMessageResponse} from '../session';
 
 export function checkEnumerate(res: mixed): Array<DeviceDescriptor> {
     if (typeof res !== 'object') {
