+++ lib/transport/chrome-extension.js
@@ -2,13 +2,12 @@
 'use strict';
 
 import semvercmp from 'semver-compare';
-
 import * as ChromeMessages from '../chrome-messages';
-import type {
-    ChromeMessage,
-    DefaultMessageResponse,
-    DeviceDescriptor,
-} from '../flowtypes';
+
+import type {DefaultMessageResponse} from '../session';
+import type {DeviceDescriptor} from '../transport';
+import type {ChromeMessage} from '../chrome-messages';
+
 import {
     checkEnumerate,
     checkAcquire,
