+++ lib/transport/http.js
@@ -11,11 +11,9 @@ import {
     checkInfo,
 } from './checks';
 
-import type {
-  RequestOptions,
-  DeviceDescriptor,
-  DefaultMessageResponse,
-} from '../flowtypes';
+import type {DefaultMessageResponse} from '../session';
+import type {RequestOptions} from '../http';
+import type {DeviceDescriptor} from '../transport';
 
 const DEFAULT_URL = 'https://localback.net:21324';
