+++ lib/transport/index.js
@@ -2,10 +2,29 @@
 import HttpTransport from './http';
 import ChromeExtensionTransport from './chrome-extension';
 import PluginTransport from './plugin';
-
 import http from '../http';
 
-import type {Transport} from '../flowtypes';
+import type {DefaultMessageResponse} from '../session';
+
+export type DeviceDescriptor = {
+    path: string;
+    session: ?string;
+    product: number;
+    serialNumber: number;
+    vendor: number;
+};
+
+export type Transport = {
+    enumerate: (wait?: boolean, previous?: ?Array<DeviceDescriptor>) => Promise<Array<DeviceDescriptor>>;
+    acquire: (i: DeviceDescriptor, checkPrevious?: boolean) => Promise<{session: string}>;
+    supportsSync: boolean;
+    release: (s: string) => Promise;
+    releaseSync: (s: string) => void;
+    configure: (s: string) => Promise;
+    call: (s: string, type: string, msg: Object) => Promise<DefaultMessageResponse>;
+    callSync: (s: string, type: string, msg: Object) => DefaultMessageResponse;
+    version: string;
+};
 
 // Attempts to load any available HW transport layer.
 export function loadTransport(): Promise<Transport> {
