+++ lib/unacquired-device.js
@@ -2,10 +2,10 @@
 'use strict';
 
 import Device from './device';
-
-import type {Transport, DeviceDescriptor} from './flowtypes';
-import type DeviceList from './device-list';
 import {EventEmitter} from './events';
+
+import type {Transport, DeviceDescriptor} from './transport';
+import type DeviceList from './device-list';
 
 export default class UnacquiredDevice extends EventEmitter {
     transport: Transport;
