+++ lib/utils/call.js
@@ -3,10 +3,11 @@
 
 import randombytes from 'randombytes';
 
-import * as types from '../flowtypes';
 import type {EventEmitter} from '../events';
+import type {DefaultMessageResponse} from '../session';
+import type {Transport} from '../transport';
 
-function assertType(res: types.DefaultMessageResponse, resType: string) {
+function assertType(res: DefaultMessageResponse, resType: string) {
     if (res.type !== resType) {
         throw new TypeError(`Response of unexpected type: ${res.type}`);
     }
@@ -43,12 +44,12 @@ function filterForLog(type: string, msg: Object): Object {
 
 export class CallHelper {
     emitter: EventEmitter;
-    transport: types.Transport;
+    transport: Transport;
     sessionId: string;
     supportsSync: boolean;
 
     constructor(
-        transport: types.Transport,
+        transport: Transport,
         emitter: EventEmitter,
         sessionId: string,
         supportsSync: boolean
@@ -60,7 +61,7 @@ export class CallHelper {
     }
 
     // Sends an async message to the opened device.
-    call(type: string, msg_: ?Object): Promise<types.DefaultMessageResponse> {
+    call(type: string, msg_: ?Object): Promise<DefaultMessageResponse> {
         const msg = msg_ == null ? {} : msg_;
         const logMessage: Object = filterForLog(type, msg);
 
@@ -68,7 +69,7 @@ export class CallHelper {
         this.emitter.emit('send', type, msg);
 
         return this.transport.call(this.sessionId, type, msg).then(
-            (res: types.DefaultMessageResponse) => {
+            (res: DefaultMessageResponse) => {
                 const logMessage = filterForLog(res.type, res.message);
 
                 console.log('[trezor] Received', res.type, logMessage);
@@ -83,7 +84,7 @@ export class CallHelper {
         );
     }
 
-    callSync(type: string, msg_: ?Object): types.DefaultMessageResponse {
+    callSync(type: string, msg_: ?Object): DefaultMessageResponse {
         const msg = msg_ == null ? {} : msg_;
         if (!this.supportsSync) {
             throw new Error('Blocking calls are not supported');
@@ -110,7 +111,7 @@ export class CallHelper {
         }
     }
 
-    typedCall(type: string, resType: string, msg_: ?Object): Promise<types.DefaultMessageResponse> {
+    typedCall(type: string, resType: string, msg_: ?Object): Promise<DefaultMessageResponse> {
         const msg = msg_ == null ? {} : msg_;
         return this._commonCall(type, msg).then(res => {
             assertType(res, resType);
@@ -118,13 +119,13 @@ export class CallHelper {
         });
     }
 
-    _commonCall(type: string, msg: Object): Promise<types.DefaultMessageResponse> {
+    _commonCall(type: string, msg: Object): Promise<DefaultMessageResponse> {
         return this.call(type, msg).then(res =>
             this._filterCommonTypes(res)
         );
     }
 
-    _filterCommonTypes(res: types.DefaultMessageResponse): Promise<types.DefaultMessageResponse> {
+    _filterCommonTypes(res: DefaultMessageResponse): Promise<DefaultMessageResponse> {
         if (res.type === 'Failure') {
             return Promise.reject(new Error(res.message.message));
         }
