+++ lib/utils/hdnode.js
@@ -4,7 +4,8 @@
 import * as bitcoin from 'bitcoinjs-lib';
 import * as ecurve from 'ecurve';
 import * as types from '../flowtypes';
-import type Session from '../session';
+
+import type Session, {MessageResponse} from '../session';
 
 // Converts from bitcoin.js HDNode to Trezor's own structure
 export function bjsNode2privNode(node: bitcoin.HDNode): types.HDPrivNode {
@@ -41,7 +42,7 @@ export function pubNode2bjsNode(
 // network info is necessary
 // throws error on wrong xpub
 export function pubKey2bjsNode(
-    key: types.MessageResponse<types.PublicKey>,
+    key: MessageResponse<types.PublicKey>,
     network: bitcoin.Network
 ): bitcoin.HDNode {
     const keyNode: types.HDPubNode = key.message.node;
@@ -97,10 +98,10 @@ export function getHDNode(
     const childPath: Array<number> = path.concat([suffix]);
     const network: bitcoin.Network = network_ == null ? bitcoin.networks.bitcoin : network_;
 
-    return session.getPublicKey(path).then((resKey: types.MessageResponse<types.PublicKey>) => {
+    return session.getPublicKey(path).then((resKey: MessageResponse<types.PublicKey>) => {
         const resNode: bitcoin.HDNode = pubKey2bjsNode(resKey, network);
 
-        return session.getPublicKey(childPath).then((childKey: types.MessageResponse<types.PublicKey>) => {
+        return session.getPublicKey(childPath).then((childKey: MessageResponse<types.PublicKey>) => {
             const childNode: bitcoin.HDNode = pubKey2bjsNode(childKey, network);
             checkDerivation(resNode, childNode, suffix);
             return resNode;
