+++ lib/utils/signbjstx.js
@@ -5,7 +5,28 @@ import * as bitcoin from 'bitcoinjs-lib';
 import * as types from '../flowtypes';
 import * as hdnodeUtils from './hdnode';
 
-function input2trezor(input: types.InputInfo): types.TransactionInput {
+import type {MessageResponse} from '../session';
+
+export type OutputInfo = {
+    path: Array<number>;
+    value: number;
+} | {
+    address: string;
+    value: number;
+};
+
+export type InputInfo = {
+    hash: Buffer;
+    index: number;
+    path?: Array<number>;
+};
+
+export type TxInfo = {
+    inputs: Array<InputInfo>;
+    outputs: Array<OutputInfo>;
+};
+
+function input2trezor(input: InputInfo): types.TransactionInput {
     const {hash, index, path} = input;
     return {
         prev_index: index,
@@ -14,7 +35,7 @@ function input2trezor(input: types.InputInfo): types.TransactionInput {
     };
 }
 
-function output2trezor(output: types.OutputInfo, network: bitcoin.Network): types.TransactionOutput {
+function output2trezor(output: OutputInfo, network: bitcoin.Network): types.TransactionOutput {
     if (output.address == null) {
         if (output.path == null) {
             throw new Error('Both address and path of an output cannot be null.');
@@ -35,7 +56,7 @@ function output2trezor(output: types.OutputInfo, network: bitcoin.Network): type
     };
 }
 
-function signedTx2refTx(signedTx: types.MessageResponse<types.SignedTx>): bitcoin.Transaction {
+function signedTx2refTx(signedTx: MessageResponse<types.SignedTx>): bitcoin.Transaction {
     const res = bitcoin.Transaction.fromHex(signedTx.message.serialized.serialized_tx);
     return res;
 }
@@ -89,8 +110,8 @@ function deriveOutputScript(
 }
 
 function verifyBjsTx(
-    inputs: Array<types.InputInfo>,
-    outputs: Array<types.OutputInfo>,
+    inputs: Array<InputInfo>,
+    outputs: Array<OutputInfo>,
     nodes: Array<bitcoin.HDNode>,
     resTx: bitcoin.Transaction,
     network: bitcoin.Network
@@ -138,7 +159,7 @@ function getAddressScriptType(address: string, network: bitcoin.Network): string
 }
 
 export function signBjsTx(
-    info: types.TxInfo,
+    info: TxInfo,
     refTxs: Array<bitcoin.Transaction>,
     nodes: Array<bitcoin.HDNode>,
     coinName: string
