+++ lib/utils/signtx.js
@@ -2,7 +2,8 @@
 'use strict';
 
 import * as types from '../flowtypes';
-import type Session from '../session';
+
+import type Session, {MessageResponse} from '../session';
 
 function indexTxsForSign(
     txs: Array<types.RefTransaction>
@@ -113,7 +114,7 @@ function processTxRequest(
     index: {[key: string]: types.RefTransaction},
     ins: Array<types.TransactionInput>,
     outs: Array<types.TransactionOutput>
-): Promise<types.MessageResponse<types.SignedTx>> {
+): Promise<MessageResponse<types.SignedTx>> {
     saveTxSignatures(m.serialized, serializedTx, signatures);
 
     if (m.request_type === 'TXFINISHED') {
@@ -140,7 +141,7 @@ export function signTx(
     outputs: Array<types.TransactionOutput>,
     txs: Array<types.RefTransaction>,
     coin: types.CoinType | string
-): Promise<types.MessageResponse<types.SignedTx>> {
+): Promise<MessageResponse<types.SignedTx>> {
     const index: {[key: string]: types.RefTransaction} = indexTxsForSign(txs);
     const signatures: Array<string> = [];
     const serializedTx: {serialized: string} = {serialized: ''};
